fix some bug

+++ src/utils/data_loader/level_loader.rs
@@ -384,7 +384,7 @@ impl Loader {
         let spawner:Spawner = level.waves.clone().into();
         let f = Frame {
             map,
-            cost: level.options.initial_cost,
+            cost: level.options.initial_cost as f32,
             ..Default::default()
         };
         let c = Calculator {
